requirements.txt
Django~=2.2.0
djangorestframework~=3.11
djangorestframework-jwt~=1.10
requests==2.23.0

# tests
pytest==5.2.1
pytest-django==3.5.1
pytest-mock==1.11.1
